Reduce code duplication in tests
- Refactored MageFunction tests into an abstract test case
- Added phpcpd to scan for copy/paste
