+++ composer.json
@@ -50,6 +50,7 @@
     },
     "require-dev": {
         "phpunit/phpunit": "4.1.0",
+        "sebastian/phpcpd": "2.0.0",
         "squizlabs/php_codesniffer": "1.5.3",
         "phpmd/phpmd": "@stable",
         "pdepend/pdepend": "2.0.6",
