+++ tests/unit/Magento/Migration/Code/Processor/Mage/MageFunction/GetSingletonTest.php
@@ -5,12 +5,10 @@
  */
 namespace Magento\Migration\Code\Processor\Mage\MageFunction;
 
-
 use Magento\Migration\Code\Processor\Mage\MageFunctionInterface;
-use Magento\Migration\Code\TestCase;
 use Magento\Migration\Mapping\Alias;
 
-class GetSingletonTest extends TestCase
+class GetSingletonTest extends AbstractMageFunctionTestCase
 {
     /**
      * @var GetSingleton
@@ -18,55 +16,15 @@ class GetSingletonTest extends TestCase
     protected $obj;
 
     /**
-     * @var \Magento\Migration\Mapping\ClassMapping|\PHPUnit_Framework_MockObject_MockObject
-     */
-    protected $classMapperMock;
-
-    /**
-     * @var \Magento\Migration\Mapping\Alias|\PHPUnit_Framework_MockObject_MockObject
-     */
-    protected $aliasMapperMock;
-
-    /**
-     * @var \Magento\Migration\Logger\Logger|\PHPUnit_Framework_MockObject_MockObject
-     */
-    protected $loggerMock;
-
-    /**
-     * @var \Magento\Migration\Code\Processor\TokenHelper
-     */
-    protected $tokenHelper;
-
-    /**
-     * @var \Magento\Migration\Code\Processor\Mage\MageFunction\ArgumentFactory|\PHPUnit_Framework_MockObject_MockObject
-     */
-    protected $argumentFactoryMock;
-
-    /**
      * @var \Magento\Migration\Code\Processor\NamingHelper|\PHPUnit_Framework_MockObject_MockObject
      */
     protected $namingHelperMock;
 
-    public function setUp()
+    /**
+     * @return GetSingleton
+     */
+    protected function getSubjectUnderTest()
     {
-        $this->loggerMock = $this->getMock('\Magento\Migration\Logger\Logger');
-
-        $this->classMapperMock = $this->getMockBuilder(
-            '\Magento\Migration\Mapping\ClassMapping'
-        )->disableOriginalConstructor()
-            ->getMock();
-        $this->aliasMapperMock = $this->getMockBuilder(
-            '\Magento\Migration\Mapping\Alias'
-        )->disableOriginalConstructor()
-            ->getMock();
-
-        $this->tokenHelper = $this->setupTokenHelper($this->loggerMock);
-
-        $this->argumentFactoryMock = $this->getMockBuilder(
-            '\Magento\Migration\Code\Processor\Mage\MageFunction\ArgumentFactory'
-        )->setMethods(['create'])
-            ->getMock();
-
         $this->namingHelperMock = $this->getMockBuilder(
             '\Magento\Migration\Code\Processor\NamingHelper'
         )->disableOriginalConstructor()
@@ -94,7 +52,7 @@ class GetSingletonTest extends TestCase
                 ['\\Magento\\Tax\\Model\\Rule', 'taxRule'],
             ]);
 
-        $this->obj = new GetSingleton(
+        return new GetSingleton(
             $this->classMapperMock,
             $this->aliasMapperMock,
             $this->loggerMock,
@@ -105,7 +63,10 @@ class GetSingletonTest extends TestCase
     }
 
     /**
-     * @dataProvider getSingletonDataProvider
+     * Execute test suite on the test subject using data provider
+     *
+     * @dataProvider dataProvider
+     *
      * @param $inputFile
      * @param $index
      * @param $attrs
@@ -120,14 +81,8 @@ class GetSingletonTest extends TestCase
         $m1ClassName,
         $mappedModelClass,
         $expectedFile
-    ) {
-        $file = __DIR__ . '/_files/' . $inputFile;
-        $fileContent = file_get_contents($file);
-
-        $tokens = token_get_all($fileContent);
-
-        $this->obj->setContext($tokens, $index);
-
+    )
+    {
         $this->aliasMapperMock->expects($this->any())
             ->method('mapAlias')
             ->with('tax', 'model')
@@ -138,24 +93,10 @@ class GetSingletonTest extends TestCase
             ->with($m1ClassName)
             ->willReturn($mappedModelClass);
 
-        $this->assertEquals($attrs['start_index'], $this->obj->getStartIndex());
-        $this->assertEquals($attrs['end_index'], $this->obj->getEndIndex());
-        $this->assertEquals($attrs['method'], $this->obj->getMethod());
-        $this->assertEquals($attrs['type'], $this->obj->getType());
-        $this->assertEquals($attrs['class'], $this->obj->getClass());
-        $this->assertEquals($attrs['di_variable_name'], $this->obj->getDiVariableName());
-        $this->assertEquals($attrs['di_variable_class'], $this->obj->getDiClass());
-
-        $this->obj->convertToM2();
-
-        $updatedContent = $this->tokenHelper->reconstructContent($tokens);
-
-        $expectedFile = __DIR__ . '/_files/' . $expectedFile;
-        $expected = file_get_contents($expectedFile);
-        $this->assertEquals($expected, $updatedContent);
+        $this->executeTestAgainstExpectedFile($inputFile, $index, $attrs, $expectedFile);
     }
 
-    public function getSingletonDataProvider()
+    public function dataProvider()
     {
         $data = [
             'singleton_mapped' => [
