+++ tests/unit/Magento/Migration/Code/Processor/Mage/MageFunction/HelperTest.php
@@ -6,11 +6,10 @@
 namespace Magento\Migration\Code\Processor\Mage\MageFunction;
 
 use Magento\Migration\Code\Processor\Mage\MageFunctionInterface;
-use Magento\Migration\Code\TestCase;
+use Magento\Migration\Code\Processor\NamingHelper;
 use Magento\Migration\Mapping\Alias;
 
-class HelperTest extends TestCase
-
+class HelperTest extends AbstractMageFunctionTestCase
 {
     /**
      * @var Helper
@@ -18,55 +17,15 @@ class HelperTest extends TestCase
     protected $obj;
 
     /**
-     * @var \Magento\Migration\Mapping\ClassMapping|\PHPUnit_Framework_MockObject_MockObject
-     */
-    protected $classMapperMock;
-
-    /**
-     * @var \Magento\Migration\Mapping\Alias|\PHPUnit_Framework_MockObject_MockObject
-     */
-    protected $aliasMapperMock;
-
-    /**
-     * @var \Magento\Migration\Logger\Logger|\PHPUnit_Framework_MockObject_MockObject
-     */
-    protected $loggerMock;
-
-    /**
-     * @var \Magento\Migration\Code\Processor\TokenHelper
-     */
-    protected $tokenHelper;
-
-    /**
-     * @var \Magento\Migration\Code\Processor\Mage\MageFunction\ArgumentFactory|\PHPUnit_Framework_MockObject_MockObject
-     */
-    protected $argumentFactoryMock;
-
-    /**
-     * @var \Magento\Migration\Code\Processor\NamingHelper|\PHPUnit_Framework_MockObject_MockObject
+     * @var NamingHelper|\PHPUnit_Framework_MockObject_MockObject
      */
     protected $namingHelperMock;
 
-    public function setUp()
+    /**
+     * @return GetStoreConfig
+     */
+    protected function getSubjectUnderTest()
     {
-        $this->loggerMock = $this->getMock('\Magento\Migration\Logger\Logger');
-
-        $this->classMapperMock = $this->getMockBuilder(
-            '\Magento\Migration\Mapping\ClassMapping'
-        )->disableOriginalConstructor()
-            ->getMock();
-        $this->aliasMapperMock = $this->getMockBuilder(
-            '\Magento\Migration\Mapping\Alias'
-        )->disableOriginalConstructor()
-            ->getMock();
-
-        $this->tokenHelper = $this->setupTokenHelper($this->loggerMock);
-
-        $this->argumentFactoryMock = $this->getMockBuilder(
-            '\Magento\Migration\Code\Processor\Mage\MageFunction\ArgumentFactory'
-        )->setMethods(['create'])
-            ->getMock();
-
         $this->namingHelperMock = $this->getMockBuilder(
             '\Magento\Migration\Code\Processor\NamingHelper'
         )->disableOriginalConstructor()
@@ -88,7 +47,7 @@ class HelperTest extends TestCase
                 ['Mage_Tax_Helper_Config', '\\Magento\\Tax\\Helper\\Config'],
             ]);
 
-        $this->obj = new Helper(
+        return new Helper(
             $this->classMapperMock,
             $this->aliasMapperMock,
             $this->loggerMock,
@@ -113,13 +72,6 @@ class HelperTest extends TestCase
         $mappedHelperClass,
         $expectedFile
     ) {
-        $file = __DIR__ . '/_files/' . $inputFile;
-        $fileContent = file_get_contents($file);
-
-        $tokens = token_get_all($fileContent);
-
-        $this->obj->setContext($tokens, $index);
-
         $this->aliasMapperMock->expects($this->any())
             ->method('mapAlias')
             ->with('tax', 'helper')
@@ -129,21 +81,7 @@ class HelperTest extends TestCase
             ->method('mapM1Class')
             ->willReturn($mappedHelperClass);
 
-        $this->assertEquals($attrs['start_index'], $this->obj->getStartIndex());
-        $this->assertEquals($attrs['end_index'], $this->obj->getEndIndex());
-        $this->assertEquals($attrs['method'], $this->obj->getMethod());
-        $this->assertEquals($attrs['type'], $this->obj->getType());
-        $this->assertEquals($attrs['class'], $this->obj->getClass());
-        $this->assertEquals($attrs['di_variable_name'], $this->obj->getDiVariableName());
-        $this->assertEquals($attrs['di_variable_class'], $this->obj->getDiClass());
-
-        $this->obj->convertToM2();
-
-        $updatedContent = $this->tokenHelper->reconstructContent($tokens);
-
-        $expectedFile = __DIR__ . '/_files/' . $expectedFile;
-        $expected = file_get_contents($expectedFile);
-        $this->assertEquals($expected, $updatedContent);
+        $this->executeTestAgainstExpectedFile($inputFile, $index, $attrs, $expectedFile);
     }
 
     public function helperDataProvider()
