+++ tests/unit/Magento/Migration/Code/Processor/Mage/MageFunction/RegistryTest.php
@@ -6,9 +6,8 @@
 namespace Magento\Migration\Code\Processor\Mage\MageFunction;
 
 use Magento\Migration\Code\Processor\Mage\MageFunctionInterface;
-use Magento\Migration\Code\TestCase;
 
-class RegistryTest extends TestCase
+class RegistryTest extends AbstractMageFunctionTestCase
 {
     /**
      * @var Registry
@@ -16,51 +15,11 @@ class RegistryTest extends TestCase
     protected $obj;
 
     /**
-     * @var \Magento\Migration\Mapping\ClassMapping|\PHPUnit_Framework_MockObject_MockObject
+     * @return Registry
      */
-    protected $classMapperMock;
-
-    /**
-     * @var \Magento\Migration\Mapping\Alias|\PHPUnit_Framework_MockObject_MockObject
-     */
-    protected $aliasMapperMock;
-
-    /**
-     * @var \Magento\Migration\Logger\Logger|\PHPUnit_Framework_MockObject_MockObject
-     */
-    protected $loggerMock;
-
-    /**
-     * @var \Magento\Migration\Code\Processor\TokenHelper
-     */
-    protected $tokenHelper;
-
-    /**
-     * @var \Magento\Migration\Code\Processor\Mage\MageFunction\ArgumentFactory|\PHPUnit_Framework_MockObject_MockObject
-     */
-    protected $argumentFactoryMock;
-
-    public function setUp()
+    protected function getSubjectUnderTest()
     {
-        $this->loggerMock = $this->getMock('\Magento\Migration\Logger\Logger');
-
-        $this->classMapperMock = $this->getMockBuilder(
-            '\Magento\Migration\Mapping\ClassMapping'
-        )->disableOriginalConstructor()
-            ->getMock();
-        $this->aliasMapperMock = $this->getMockBuilder(
-            '\Magento\Migration\Mapping\Alias'
-        )->disableOriginalConstructor()
-            ->getMock();
-
-        $this->tokenHelper = $this->setupTokenHelper($this->loggerMock);
-
-        $this->argumentFactoryMock = $this->getMockBuilder(
-            '\Magento\Migration\Code\Processor\Mage\MageFunction\ArgumentFactory'
-        )->setMethods(['create'])
-            ->getMock();
-
-        $this->obj = new Registry(
+        return new Registry(
             $this->classMapperMock,
             $this->aliasMapperMock,
             $this->loggerMock,
@@ -82,28 +41,7 @@ class RegistryTest extends TestCase
         $attrs,
         $expectedFile
     ) {
-        $file = __DIR__ . '/_files/' . $inputFile;
-        $fileContent = file_get_contents($file);
-
-        $tokens = token_get_all($fileContent);
-
-        $this->obj->setContext($tokens, $index);
-
-        $this->assertEquals($attrs['start_index'], $this->obj->getStartIndex());
-        $this->assertEquals($attrs['end_index'], $this->obj->getEndIndex());
-        $this->assertEquals($attrs['method'], $this->obj->getMethod());
-        $this->assertEquals($attrs['type'], $this->obj->getType());
-        $this->assertEquals($attrs['class'], $this->obj->getClass());
-        $this->assertEquals($attrs['di_variable_name'], $this->obj->getDiVariableName());
-        $this->assertEquals($attrs['di_variable_class'], $this->obj->getDiClass());
-
-        $this->obj->convertToM2();
-
-        $updatedContent = $this->tokenHelper->reconstructContent($tokens);
-
-        $expectedFile = __DIR__ . '/_files/' . $expectedFile;
-        $expected = file_get_contents($expectedFile);
-        $this->assertEquals($expected, $updatedContent);
+        $this->executeTestAgainstExpectedFile($inputFile, $index, $attrs, $expectedFile);
     }
 
     public function registryDataProvider()
